Queue and Resolution management
Resolution settings up to 512 channels - works with the new command gen
Queue management:
- barriers on the ends
- new "readable_size" parameter in queue struct (to not let it overflow and solve flash problems, saving was bad)
globalized some variables (i think)
commented "queue_delete" function, that we havent used a long time ago

-notice: flash only saves 10 items? needs fix

+++ Core/Inc/Request.h
@@ -18,6 +18,7 @@ typedef enum class {
 	MAX_HITS = 0x03
 } RequestType;
 
+
 typedef struct {
 	uint8_t ID;				// (new) measurement ID
 	RequestType type;
@@ -29,7 +30,7 @@ typedef struct {
 	uint16_t min_voltage;
 	uint16_t max_voltage;
 	uint8_t samples;
-	uint8_t resolution;
+	uint16_t resolution;
 } Request;
 
 static Request empty_request = {
